fix(emoji): 🐛 remount the card template after flip animation on android
On android when a card has a big emoji (see the flags deck) after a flip animation the emoji would
become blurred and then keep getting blurrier after more flips. This fix remounts the template after
a flip, which fixes the issue at the expense of remounting. However this only happens on specific
user gestures so it's not so bad for now. Will want to fix the actual blurring at some point but
this may require our own custom emoji font or something. Perhaps even emojis as icons or images

branch dev